CW-83 Move events into a separate package
