+++ backend/server/events/events.go
@@ -1,28 +1,28 @@
-package main
+package events
 
 import (
-	"time"
-	"fmt"
 	"bytes"
 	"encoding/json"
+	"fmt"
 	"sort"
-	// "os"
-	// "log"
-	"path/filepath"
+	"time"
+
 	"io/ioutil"
 	"net/http"
-	. "csesoc.unsw.edu.au/m/v2/server"
-	"github.com/relvacode/iso8601"
-	"github.com/labstack/echo/v4"
-)
+	"path/filepath"
 
+	. "csesoc.unsw.edu.au/m/v2/server"
+
+	"github.com/labstack/echo/v4"
+	"github.com/relvacode/iso8601"
+)
 
 // FB response expects data (array of events), and paging, which we ignore.
 // However, should FB provide an error, we capture it.
 type FbResponse struct {
 	Data  []FbRespEvent `json:"data"`
 	Cover FbRespCover   `json:"cover"`
-	Error FbRespError   `json:"error"` 
+	Error FbRespError   `json:"error"`
 }
 
 // Unmarshal event specifics.
@@ -75,35 +75,35 @@ type Event struct {
 }
 
 // Sets up a ticker to fetch events at an interval.
-func eventFetchTimer() {
+func EventFetchTimer() {
 	for range time.Tick(time.Duration(FB_FETCH_INTERVAL * time.Second)) {
 		saveEvents()
 	}
 }
 
 // Fetch events from FB
-func fetchEvents(response *FbResponse) (error) {
+func fetchEvents(response *FbResponse) error {
 	// Make a request to FB
 	resp, err := http.Get(
 		fmt.Sprintf(
 			"%s%s?access_token=%s&since=%d",
-			FB_API_PATH, 
+			FB_API_PATH,
 			FB_EVENT_PATH,
 			FB_TOKEN,
 			time.Now().Unix()),
-		)
+	)
 	if err != nil {
 		return fmt.Errorf("There was an error making a request to FB")
 	}
 	defer resp.Body.Close()
 
 	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("There was an error making a request to FB")	
+		return fmt.Errorf("There was an error making a request to FB")
 	}
 	body, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
 		return fmt.Errorf("Something went wrong with parsing the FB request")
-	} 
+	}
 
 	json.Unmarshal(body, &response)
 
@@ -116,119 +116,119 @@ func fetchEvents(response *FbResponse) (error) {
 
 // Fetch the cover image for an event from Facebook.
 func fetchCoverImage(id string) (string, error) {
-		resp, err := http.Get(
-			fmt.Sprintf(
-				"%s/%s?fields=cover&access_token=%s",
-				FB_API_PATH,
-				id,
-				FB_TOKEN,
-			),
-		)
-		if err != nil {
-			return "", fmt.Errorf("There was an error making a request to FB")
-		}
+	resp, err := http.Get(
+		fmt.Sprintf(
+			"%s/%s?fields=cover&access_token=%s",
+			FB_API_PATH,
+			id,
+			FB_TOKEN,
+		),
+	)
+	if err != nil {
+		return "", fmt.Errorf("There was an error making a request to FB")
+	}
 
-		if resp.StatusCode != http.StatusOK {
-			return "", fmt.Errorf("There was an error making a request to FB.")	
-		}
-		body, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			return "", fmt.Errorf("Something went wrong with parsing the FB request.")
-		}
+	if resp.StatusCode != http.StatusOK {
+		return "", fmt.Errorf("There was an error making a request to FB.")
+	}
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", fmt.Errorf("Something went wrong with parsing the FB request.")
+	}
 
-		// Unmarshal the response body to pull the cover image from it
-		var result FbResponse
-		err = json.Unmarshal(body, &result)
-		if err != nil {
-			return "", fmt.Errorf("There was an issue parsing JSON.")
-		}
-		if result.Error != (FbRespError{}) {
-			return "", fmt.Errorf("Something went wrong with the FB request.")
-		}
-		return result.Cover.CoverUri, nil
+	// Unmarshal the response body to pull the cover image from it
+	var result FbResponse
+	err = json.Unmarshal(body, &result)
+	if err != nil {
+		return "", fmt.Errorf("There was an issue parsing JSON.")
+	}
+	if result.Error != (FbRespError{}) {
+		return "", fmt.Errorf("Something went wrong with the FB request.")
+	}
+	return result.Cover.CoverUri, nil
 }
 
 // Process the facebook event information before saving it to a new file.
 func saveEvents() {
-		var result FbResponse
-		err := fetchEvents(&result)
-		if err != nil {
-			// do something
-		}
+	var result FbResponse
+	err := fetchEvents(&result)
+	if err != nil {
+		// do something
+	}
 
-		// Store processed events
-		var processedEvents []Event
+	// Store processed events
+	var processedEvents []Event
 
-		for _, element := range result.Data {
-			if len(element.EventTimes) != 0 {
-				for _, occurrence := range element.EventTimes {
-					cover, err := fetchCoverImage(element.Id)
-					if err != nil {
-						// do something
-					}
-					start, err := iso8601.ParseString(occurrence.Start)
-					if err != nil {
-						// do something
-					}
-					end, err := iso8601.ParseString(occurrence.End) 
-					if err != nil {
-						// do something
-					}
-					
-					processedEvents = append(processedEvents, Event {
-						Name: element.Name,
-						Description: element.Description,
-						Start: start.Unix(),
-						End: end.Unix(),
-						Id: element.Id,
-						CoverUrl: cover,
-					})
-				}
-			} else {
+	for _, element := range result.Data {
+		if len(element.EventTimes) != 0 {
+			for _, occurrence := range element.EventTimes {
 				cover, err := fetchCoverImage(element.Id)
 				if err != nil {
 					// do something
 				}
-				start, err := iso8601.ParseString(element.Start)
+				start, err := iso8601.ParseString(occurrence.Start)
 				if err != nil {
 					// do something
 				}
-				end, err := iso8601.ParseString(element.End) 
+				end, err := iso8601.ParseString(occurrence.End)
 				if err != nil {
 					// do something
 				}
-				
-				processedEvents = append(processedEvents, Event {
-					Name: element.Name,
+
+				processedEvents = append(processedEvents, Event{
+					Name:        element.Name,
 					Description: element.Description,
-					Start: start.Unix(),
-					End: end.Unix(),
-					Id: element.Id,
-					CoverUrl: cover,
+					Start:       start.Unix(),
+					End:         end.Unix(),
+					Id:          element.Id,
+					CoverUrl:    cover,
 				})
-			}			
+			}
+		} else {
+			cover, err := fetchCoverImage(element.Id)
+			if err != nil {
+				// do something
+			}
+			start, err := iso8601.ParseString(element.Start)
+			if err != nil {
+				// do something
+			}
+			end, err := iso8601.ParseString(element.End)
+			if err != nil {
+				// do something
+			}
+
+			processedEvents = append(processedEvents, Event{
+				Name:        element.Name,
+				Description: element.Description,
+				Start:       start.Unix(),
+				End:         end.Unix(),
+				Id:          element.Id,
+				CoverUrl:    cover,
+			})
 		}
+	}
 
-		// Sort by starting dates:
-		sort.Slice(processedEvents, func(i, j int) bool {
-			return processedEvents[i].Start < processedEvents[j].Start
-		 })
+	// Sort by starting dates:
+	sort.Slice(processedEvents, func(i, j int) bool {
+		return processedEvents[i].Start < processedEvents[j].Start
+	})
 
-		buf := bytes.NewBuffer([]byte{})
-		// Avoids escaping ampersands:
-		jsonEncoder := json.NewEncoder(buf)
-		jsonEncoder.SetEscapeHTML(false)
-		jsonEncoder.Encode(MarshalledEvents {
-			LastUpdate: time.Now().Unix(),
-			Events: processedEvents,
-		})
-		fp, _ := filepath.Abs("static/events.json")
-		// 644 = chmod a+rwx,u-x,g-wx,o-wx
-		err = ioutil.WriteFile(fp, buf.Bytes(), 0644)
+	buf := bytes.NewBuffer([]byte{})
+	// Avoids escaping ampersands:
+	jsonEncoder := json.NewEncoder(buf)
+	jsonEncoder.SetEscapeHTML(false)
+	jsonEncoder.Encode(MarshalledEvents{
+		LastUpdate: time.Now().Unix(),
+		Events:     processedEvents,
+	})
+	fp, _ := filepath.Abs("static/events.json")
+	// 644 = chmod a+rwx,u-x,g-wx,o-wx
+	err = ioutil.WriteFile(fp, buf.Bytes(), 0644)
 
-		if err != nil {
-			// error handling
-		}				
+	if err != nil {
+		// error handling
+	}
 }
 
 func GetEvents(c echo.Context) error {
